+++ Products/ATContentTypes/exportimport/content.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from Products.GenericSetup.interfaces import IFilesystemExporter
-from zope.interface import implements
+from zope.interface import implementer
 from zope.interface import Interface
 
 
@@ -12,11 +12,10 @@ class IDisabledExport(Interface):
     pass
 
 
+@implementer(IFilesystemExporter)
 class NullExporterAdapter(object):
     """Dummy exporter that does nothing
     """
-
-    implements(IFilesystemExporter)
 
     def __init__(self, context):
         self.context = context